rename: CoffeeStackEmptyModal 네이밍 수정
- CoffeeStackModal로 변경
