--- frontend/src/components/CoffeeStackEmptyModal/CoffeeStackEmptyModal.stories.jsx
@@ -1,13 +1,0 @@
-import CoffeeStackEmptyModal from '.';
-
-export default {
-  title: 'Components/CoffeeStackEmptyModal',
-  component: CoffeeStackEmptyModal,
-};
-
-const Template = (args) => {
-  return <CoffeeStackEmptyModal {...args} />;
-};
-
-export const Default = Template.bind({});
-Default.args = {};
