+++ frontend/src/components/CoffeeStackModal/CoffeeStackModal.stories.jsx
@@ -1,0 +1,13 @@
+import CoffeeStackModal from '.';
+
+export default {
+  title: 'Components/CoffeeStackModal',
+  component: CoffeeStackModal,
+};
+
+const Template = (args) => {
+  return <CoffeeStackModal {...args} />;
+};
+
+export const Default = Template.bind({});
+Default.args = {};
